requirements.txt
imageio==2.21.3
matplotlib==2.2.2
numpy==1.16
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.3
scikit-image==0.14.2
scikit-learn==0.19.2
scipy==1.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.24.0
imageio-ffmpeg==0.4.7
opencv-python
face_alignment